messaging_app/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
drf-nested-routers==0.95.0
djangorestframework-simplejwt==5.3.1
